feat: implement upload proof dialog/bottomsheet
